Split TrainerPopup into a create and edit variant
