+++ src/com/company/popup/TrainerPopup.java
@@ -6,9 +6,9 @@ import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 
-public class TrainerPopup extends Popup<Trainer>{
-    private TextField trainer;
-    private TextField pokemon;
+public abstract class TrainerPopup extends Popup<Trainer>{
+    protected TextField trainer;
+    protected TextField pokemon;
 
     public TrainerPopup(Trainer item) {
         super(item);
@@ -24,27 +24,11 @@ public class TrainerPopup extends Popup<Trainer>{
         HBox pokemonBox = new HBox(label2, pokemon);
 
         setInputArea(new VBox(trainerBox, pokemonBox));
-
-        if (item == null)
-            return;
-        trainer.setText(item.getName());
-        pokemon.setText(item.getPokemon());
     }
 
     @Override
-    public void apply() {
-        if (item == null) {
-            new Trainer(trainer.getText(), pokemon.getText());
-        } else {
-            item.setName(trainer.getText());
-            item.setPokemon(pokemon.getText());
-        }
-    }
+    public abstract void apply();
 
     @Override
-    public void delete() {
-        if (item != null) {
-            item.remove();
-        }
-    }
+    public abstract void delete();
 }
